"Add NextDns Linked IP updater service"
This commit introduces a new NextDns Linked IP updater service. The service is designed to run as a background task, performing updates every 5 minutes. The service is containerized using Docker and integrated with Balena for device management. It's built using .NET 7.0 and it uses Flurl.Http for HTTP requests. The updater service can be controlled and monitored utilizing a newly added LinkedIp controller, which exposes an HTTP Get endpoint for manually triggering an update. This is essential for keeping NextDns' linked IP feature updated with the dynamic IP address assigned by the internet service provider.